pyproject.toml
[tool.poetry]
name = "hw-drf"
version = "0.1.0"
description = ""
authors = ["Denis_L <denis88lyapin@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_drf"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
psycopg2-binary = "^2.9.7"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
